Move parsing to `MissionModelParser`
